// MovieStore/movie-store-app/src/app/api/api-reference.ts
export class Customer implements ICustomer {
    id?: string;
    email?: Email;
    status?: CustomerStatus;
    role?: Role;
    purchasedMovies?: PurchasedMovie[];
    moneySpent?: Money;

    constructor(data?: ICustomer) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.email = _data["email"] ? Email.fromJS(_data["email"]) : <any>undefined;
            this.status = _data["status"] ? CustomerStatus.fromJS(_data["status"]) : <any>undefined;
            this.role = _data["role"];
            if (Array.isArray(_data["purchasedMovies"])) {
                this.purchasedMovies = [] as any;
                for (let item of _data["purchasedMovies"])
                    this.purchasedMovies!.push(PurchasedMovie.fromJS(item));
            }
            this.moneySpent = _data["moneySpent"] ? Money.fromJS(_data["moneySpent"]) : <any>undefined;
        }
    }

    static fromJS(data: any): Customer {
        data = typeof data === 'object' ? data : {};
        let result = new Customer();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["email"] = this.email ? this.email.toJSON() : <any>undefined;
        data["status"] = this.status ? this.status.toJSON() : <any>undefined;
        data["role"] = this.role;
        if (Array.isArray(this.purchasedMovies)) {
            data["purchasedMovies"] = [];
            for (let item of this.purchasedMovies)
                data["purchasedMovies"].push(item.toJSON());
        }
        data["moneySpent"] = this.moneySpent ? this.moneySpent.toJSON() : <any>undefined;
        return data;
    }
}

export interface ICustomer {
    id?: string;
    email?: Email;
    status?: CustomerStatus;
    role?: Role;
    purchasedMovies?: PurchasedMovie[];
    moneySpent?: Money;
}

export class Email implements IEmail {
    value?: string;

    constructor(data?: IEmail) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.value = _data["value"];
        }
    }

    static fromJS(data: any): Email {
        data = typeof data === 'object' ? data : {};
        let result = new Email();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["value"] = this.value;
        return data;
    }
}

export interface IEmail {
    value?: string;
}

export class CustomerStatus implements ICustomerStatus {
    statusValue?: Status;
    statusExpirationDateValue?: ExpirationDate;

    constructor(data?: ICustomerStatus) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.statusValue = _data["statusValue"];
            this.statusExpirationDateValue = _data["statusExpirationDateValue"] ? ExpirationDate.fromJS(_data["statusExpirationDateValue"]) : <any>undefined;
        }
    }

    static fromJS(data: any): CustomerStatus {
        data = typeof data === 'object' ? data : {};
        let result = new CustomerStatus();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["statusValue"] = this.statusValue;
        data["statusExpirationDateValue"] = this.statusExpirationDateValue ? this.statusExpirationDateValue.toJSON() : <any>undefined;
        return data;
    }
}

export interface ICustomerStatus {
    statusValue?: Status;
    statusExpirationDateValue?: ExpirationDate;
}

export enum Status {
    Regular = 0,
    Advanced = 1,
}

export class ExpirationDate implements IExpirationDate {
    value?: Date | undefined;

    constructor(data?: IExpirationDate) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.value = _data["value"] ? new Date(_data["value"].toString()) : <any>undefined;
        }
    }

    static fromJS(data: any): ExpirationDate {
        data = typeof data === 'object' ? data : {};
        let result = new ExpirationDate();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["value"] = this.value ? this.value.toISOString() : <any>undefined;
        return data;
    }
}

export interface IExpirationDate {
    value?: Date | undefined;
}

export enum Role {
    Regular = 0,
    Administrator = 1,
}

export class PurchasedMovie implements IPurchasedMovie {
    id?: string;
    movie?: Movie | undefined;
    customer?: Customer | undefined;
    dateOfPurchase?: Date;
    movieExpirationDate?: Date | undefined;
    price?: Money | undefined;

    constructor(data?: IPurchasedMovie) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.movie = _data["movie"] ? Movie.fromJS(_data["movie"]) : <any>undefined;
            this.customer = _data["customer"] ? Customer.fromJS(_data["customer"]) : <any>undefined;
            this.dateOfPurchase = _data["dateOfPurchase"] ? new Date(_data["dateOfPurchase"].toString()) : <any>undefined;
            this.movieExpirationDate = _data["movieExpirationDate"] ? new Date(_data["movieExpirationDate"].toString()) : <any>undefined;
            this.price = _data["price"] ? Money.fromJS(_data["price"]) : <any>undefined;
        }
    }

    static fromJS(data: any): PurchasedMovie {
        data = typeof data === 'object' ? data : {};
        let result = new PurchasedMovie();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["movie"] = this.movie ? this.movie.toJSON() : <any>undefined;
        data["customer"] = this.customer ? this.customer.toJSON() : <any>undefined;
        data["dateOfPurchase"] = this.dateOfPurchase ? this.dateOfPurchase.toISOString() : <any>undefined;
        data["movieExpirationDate"] = this.movieExpirationDate ? this.movieExpirationDate.toISOString() : <any>undefined;
        data["price"] = this.price ? this.price.toJSON() : <any>undefined;
        return data;
    }
}

export interface IPurchasedMovie {
    id?: string;
    movie?: Movie | undefined;
    customer?: Customer | undefined;
    dateOfPurchase?: Date;
    movieExpirationDate?: Date | undefined;
    price?: Money | undefined;
}
// ...
export class Money implements IMoney {
    value?: number;

    constructor(data?: IMoney) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.value = _data["value"];
        }
    }

    static fromJS(data: any): Money {
        data = typeof data === 'object' ? data : {};
        let result = new Money();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["value"] = this.value;
        return data;
    }
}

export interface IMoney {
    value?: number;
}
// ...
export class UpdateCustomerCommand implements IUpdateCustomerCommand {
    id?: string;
    email?: string;
    role?: Role;

    constructor(data?: IUpdateCustomerCommand) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.email = _data["email"];
            this.role = _data["role"];
        }
    }

    static fromJS(data: any): UpdateCustomerCommand {
        data = typeof data === 'object' ? data : {};
        let result = new UpdateCustomerCommand();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["email"] = this.email;
        data["role"] = this.role;
        return data;
    }
}

export interface IUpdateCustomerCommand {
    id?: string;
    email?: string;
    role?: Role;
}

export class DeleteCustomerByIdCommand implements IDeleteCustomerByIdCommand {
    id?: string;

    constructor(data?: IDeleteCustomerByIdCommand) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
        }
    }

    static fromJS(data: any): DeleteCustomerByIdCommand {
        data = typeof data === 'object' ? data : {};
        let result = new DeleteCustomerByIdCommand();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        return data;
    }
}

export interface IDeleteCustomerByIdCommand {
    id?: string;
}
// ...
export class PromoteCustomerCommand implements IPromoteCustomerCommand {
    customerId?: string;

    constructor(data?: IPromoteCustomerCommand) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.customerId = _data["customerId"];
        }
    }

    static fromJS(data: any): PromoteCustomerCommand {
        data = typeof data === 'object' ? data : {};
        let result = new PromoteCustomerCommand();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["customerId"] = this.customerId;
        return data;
    }
}

export interface IPromoteCustomerCommand {
    customerId?: string;
}